WIP move some functions to prospect_api Update control vars full_test.py
