+++ unhcr/db.py
@@ -399,7 +399,37 @@ def update_prospect(start_ts=None, local=True):
 
     logging.info(f"Starting update_prospect ts: {start_ts}  local = {local}")
     try:
-        prospect_get_key(api_prospect.get_prospect_url_key, local, start_ts)
+        start_ts = api_prospect.prospect_get_start_ts(local, start_ts)
+        res = mysql_execute(
+            f"select * FROM defaultdb.TAKUM_LEONICS_API_RAW where DatetimeServer > '{start_ts}' order by DatetimeServer",
+            mysql_engine,
+            # {'ts':start_ts}
+        )
+        # Fetch all results as a list of dictionaries
+        rows = res.fetchall()
+
+        # Convert the result to a Pandas DataFrame
+        columns = res.keys()  # Get column names
+        df = pd.DataFrame(rows, columns=columns)
+
+        df["external_id"] = df["external_id"].astype(str).apply(lambda x: "py_" + x)
+
+        res = api_prospect.api_in_prospect(df, local)
+        if res is None:
+            logging.error("Prospect API failed, exiting")
+            exit()
+        logging.info(f"{res.status_code}:  {res.text}")
+
+        # Save the DataFrame to a CSV file
+        logger = logging.getLogger()
+        if logger.getEffectiveLevel() < logging.INFO:
+            sts = start_ts.replace(" ", "_").replace(":", "HM")
+            sts += str(local)
+            df.to_csv(f"py_pros_{sts}.csv", index=False)
+            # df.to_json(f'py_pros_{sts}.json', index=False)
+
+        logging.info("Data has been saved to 'py_pros'")
+
     except Exception as e:
         logging.error(f"PROSPECT Error occurred: {e}")
 
@@ -499,105 +529,6 @@ def prospect_backfill_key(func, local, start_ts):
         # df.to_json(f'py_pros_{sts}.json', index=False)
 
     logging.info(f"Data has been saved to 'py_pros'   LOCAL: {local}")
-
-
-def get_prospect_last_data(response):
-    """
-    Retrieves the latest timestamp from the Prospect API response.
-
-    This function takes a Prospect API response, parses it, and returns the latest timestamp
-    as a string in the format 'YYYY-MM-DD HH:MM:SS'.
-
-    Args:
-        response (requests.Response): The Prospect API response.
-
-    Returns:
-        str: The latest timestamp.
-
-    """
-
-    j = json.loads(response.text)
-    # json.dumps(j, indent=2)
-    # logging.info(f'\n\n{j['data'][0]}')
-    res = ""
-    idd = ""
-    for d in j["data"]:
-        if d["custom"]["DatetimeServer"] > res:
-            res = d["custom"]["DatetimeServer"]
-        if d["external_id"] > idd:
-            idd = d["external_id"]
-    return res
-
-
-def prospect_get_key(func, local, start_ts=None):
-    """
-    Retrieves data from the Prospect API and updates the MySQL database.
-
-    This function constructs a URL and fetches data from the Prospect API using the provided
-    function to get the necessary URL and API key. It then retrieves the latest timestamp
-    from the API response, queries the MySQL database for newer records, and sends this data
-    back to the Prospect API. If the API call fails, it logs an error and exits the program.
-
-    Args:
-        func (callable): A function that returns the API URL and key based on the 'local' flag.
-        local (bool): A flag indicating whether to retrieve data from the local or external
-                      Prospect API. When True, retrieves from the local API.
-        start_ts (str, optional): The timestamp to start retrieval from. If not provided (default),
-                                  retrieves the latest timestamp from the Prospect API.
-
-    Raises:
-        SystemExit: Exits the program if the Prospect API call fails.
-
-    Logs:
-        Various debug and informational logs, including headers, keys, URLs, and response
-        statuses. Also logs errors if API calls or database operations fail.
-    """
-    url, key = func(local, out=True)
-    sid = 1 if local else 421
-    url += f"/v1/out/custom/?size=50&page=1&q[source_id_eq]={sid}&q[s]=created_at+desc"
-    payload = {}
-    headers = {
-        "Authorization": f"Bearer {key}",
-    }
-
-    response = requests.request(
-        "GET", url, headers=headers, data=payload, verify=const.VERIFY
-    )
-    if start_ts is None:
-        start_ts = get_prospect_last_data(response)
-    j = json.loads(response.text)
-    # json.dumps(j, indent=2)
-    logging.info(f"\n\n{key}\n{url}\n{start_ts}")
-
-    res = mysql_execute(
-        f"select * FROM defaultdb.TAKUM_LEONICS_API_RAW where DatetimeServer > '{start_ts}' order by DatetimeServer",
-        mysql_engine,
-        # {'ts':start_ts}
-    )
-    # Fetch all results as a list of dictionaries
-    rows = res.fetchall()
-
-    # Convert the result to a Pandas DataFrame
-    columns = res.keys()  # Get column names
-    df = pd.DataFrame(rows, columns=columns)
-
-    df["external_id"] = df["external_id"].astype(str).apply(lambda x: "py_" + x)
-
-    res = api_prospect.api_in_prospect(df, local)
-    if res is None:
-        logging.error("Prospect API failed, exiting")
-        exit()
-    logging.info(f"{res.status_code}:  {res.text}")
-
-    # Save the DataFrame to a CSV file
-    logger = logging.getLogger()
-    if logger.getEffectiveLevel() < logging.INFO:
-        sts = start_ts.replace(" ", "_").replace(":", "HM")
-        sts += str(local)
-        df.to_csv(f"py_pros_{sts}.csv", index=False)
-        # df.to_json(f'py_pros_{sts}.json', index=False)
-
-    logging.info("Data has been saved to 'py_pros'")
 
 
 ###################################################
